separate customer service workflow from helper functions
