+++ openai_agents/workflows/customer_service_workflow.py
@@ -9,145 +9,23 @@ with workflow.unsafe.imports_passed_through():
         ItemHelpers,
         MessageOutputItem,
         RunConfig,
-        RunContextWrapper,
         Runner,
         ToolCallItem,
         ToolCallOutputItem,
         TResponseInputItem,
-        function_tool,
-        handoff,
         trace,
     )
-    from agents.extensions.handoff_prompt import RECOMMENDED_PROMPT_PREFIX
-    from pydantic import BaseModel
 
-
-### CONTEXT
-
-
-class AirlineAgentContext(BaseModel):
-    passenger_name: str | None = None
-    confirmation_number: str | None = None
-    seat_number: str | None = None
-    flight_number: str | None = None
-
-
-### TOOLS
-
-
-@function_tool(
-    name_override="faq_lookup_tool",
-    description_override="Lookup frequently asked questions.",
-)
-async def faq_lookup_tool(question: str) -> str:
-    if "bag" in question or "baggage" in question:
-        return (
-            "You are allowed to bring one bag on the plane. "
-            "It must be under 50 pounds and 22 inches x 14 inches x 9 inches."
-        )
-    elif "seats" in question or "plane" in question:
-        return (
-            "There are 120 seats on the plane. "
-            "There are 22 business class seats and 98 economy seats. "
-            "Exit rows are rows 4 and 16. "
-            "Rows 5-8 are Economy Plus, with extra legroom. "
-        )
-    elif "wifi" in question:
-        return "We have free wifi on the plane, join Airline-Wifi"
-    return "I'm sorry, I don't know the answer to that question."
-
-
-@function_tool
-async def update_seat(
-    context: RunContextWrapper[AirlineAgentContext],
-    confirmation_number: str,
-    new_seat: str,
-) -> str:
-    """
-    Update the seat for a given confirmation number.
-
-    Args:
-        confirmation_number: The confirmation number for the flight.
-        new_seat: The new seat to update to.
-    """
-    # Update the context based on the customer's input
-    context.context.confirmation_number = confirmation_number
-    context.context.seat_number = new_seat
-    # Ensure that the flight number has been set by the incoming handoff
-    assert context.context.flight_number is not None, "Flight number is required"
-    return f"Updated seat to {new_seat} for confirmation number {confirmation_number}"
-
-
-### HOOKS
-
-
-async def on_seat_booking_handoff(
-    context: RunContextWrapper[AirlineAgentContext],
-) -> None:
-    flight_number = f"FLT-{workflow.random().randint(100, 999)}"
-    context.context.flight_number = flight_number
-
-
-### AGENTS
-
-
-def init_agents() -> Agent[AirlineAgentContext]:
-    """
-    Initialize the agents for the airline customer service workflow.
-    :return: triage agent
-    """
-    faq_agent = Agent[AirlineAgentContext](
-        name="FAQ Agent",
-        handoff_description="A helpful agent that can answer questions about the airline.",
-        instructions=f"""{RECOMMENDED_PROMPT_PREFIX}
-        You are an FAQ agent. If you are speaking to a customer, you probably were transferred to from the triage agent.
-        Use the following routine to support the customer.
-        # Routine
-        1. Identify the last question asked by the customer.
-        2. Use the faq lookup tool to answer the question. Do not rely on your own knowledge.
-        3. If you cannot answer the question, transfer back to the triage agent.""",
-        tools=[faq_lookup_tool],
+    from openai_agents.workflows.customer_service import (
+        AirlineAgentContext,
+        ProcessUserMessageInput,
+        init_agents,
     )
-
-    seat_booking_agent = Agent[AirlineAgentContext](
-        name="Seat Booking Agent",
-        handoff_description="A helpful agent that can update a seat on a flight.",
-        instructions=f"""{RECOMMENDED_PROMPT_PREFIX}
-        You are a seat booking agent. If you are speaking to a customer, you probably were transferred to from the triage agent.
-        Use the following routine to support the customer.
-        # Routine
-        1. Ask for their confirmation number.
-        2. Ask the customer what their desired seat number is.
-        3. Use the update seat tool to update the seat on the flight.
-        If the customer asks a question that is not related to the routine, transfer back to the triage agent. """,
-        tools=[update_seat],
-    )
-
-    triage_agent = Agent[AirlineAgentContext](
-        name="Triage Agent",
-        handoff_description="A triage agent that can delegate a customer's request to the appropriate agent.",
-        instructions=(
-            f"{RECOMMENDED_PROMPT_PREFIX} "
-            "You are a helpful triaging agent. You can use your tools to delegate questions to other appropriate agents."
-        ),
-        handoffs=[
-            faq_agent,
-            handoff(agent=seat_booking_agent, on_handoff=on_seat_booking_handoff),
-        ],
-    )
-
-    faq_agent.handoffs.append(triage_agent)
-    seat_booking_agent.handoffs.append(triage_agent)
-    return triage_agent
-
-
-class ProcessUserMessageInput(BaseModel):
-    user_input: str
-    chat_length: int
 
 
 @workflow.defn
 class CustomerServiceWorkflow:
+    @workflow.init
     def __init__(self, input_items: list[TResponseInputItem] | None = None):
         self.run_config = RunConfig()
         self.chat_history: list[str] = []
